Add unordered option for ActiveRecord results
Adds an `unordered` option to the Response#records method to skip the
expensive step of reordering the set of ActiveRecords to match that of
the elasticsearch results.  That reordering step requires loading all
models in the result set, which can be much more expensive than just
lazy-loading a set of unordered results.

The default behavior remains unchanged.